ocr_demo/requirements.txt
paddlepaddle==2.6.2
paddleocr>=2.6.0.3
openpyxl==3.0.9
pandas==1.3.4
Pillow==9.0.0
numpy==1.26.4
opencv-python==4.7.0.72
pytesseract==0.3.10
xlrd==2.0.1
pyinstaller==6.13.0
tqdm>=4.64.0